chore: almost there with parity
Only missing the menu search index is sometimes broken
