+++ .gitignore
@@ -13,3 +13,5 @@ flamegraph.svg
 .DS_Store
 
 **.tmp.*
+
+.idea/*
